I have made no progress on the H2H part, however I've moved all the shared state up to App.js and passed it back down as props so now H2H can receive them
