+++ src/components/H2H.js
@@ -1,0 +1,2 @@
+import React, { useEffect, useState } from 'react';
+import managerData from '../constants/manager_details.json';
